backend/foodgram/requirements.txt
django-filter==21.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
asgiref==3.6.0
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Pillow==9.5.0
pycparser==2.21
requests==2.29.0
sqlparse==0.4.4
python-dotenv 